resolves #40 added basic skeleton classes and a main
