--- beaconchain/src/main/java/net/consensys/beaconchain/BlockTree.java
@@ -1,16 +1,0 @@
-package net.consensys.blocktree;
-
-import net.consensys.beaconchain.datastructures.Block;
-
-import org.web3j.abi.datatypes.generated.Int64;
-
-public final class BlockTree {
-
-  private Block head;
-  private Int64 height;
-
-  public BlockTree() {
-
-  }
-
-}
