+++ beaconchain/src/main/java/net/consensys/beaconchain/controllers/BlockchainController.java
@@ -1,0 +1,21 @@
+package net.consensys.beaconchain.controllers;
+
+public class BlockchainController {
+
+    public BlockchainController(){
+
+    }
+
+    public void init(){
+        // instatiate beaconchain
+        // register event listeners
+    }
+
+    public void start(){
+        // Begin block processing
+    }
+
+    public void stop(){
+
+    }
+}
