+++ beaconchain/src/main/java/net/consensys/beaconchain/controllers/NodeController.java
@@ -1,0 +1,33 @@
+package net.consensys.beaconchain.controllers;
+
+
+public class NodeController {
+
+    public NodeController(){
+
+    }
+
+    public void init(){
+        // initialize/register all services
+
+        // PoWchain Service
+
+        // Blockchain Service
+
+        // Validator Service
+
+        // P2P Service
+
+        // RPC Service
+    }
+
+    public void start(){
+        // start all services
+
+
+    }
+
+    public void stop(){
+        // stop all services
+    }
+}
