+++ beaconchain/src/main/java/net/consensys/beaconchain/controllers/PowchainController.java
@@ -1,0 +1,22 @@
+package net.consensys.beaconchain.controllers;
+
+
+public class PowchainController {
+
+    public PowchainController(){
+
+    }
+
+    public void init(){
+        // instatiate VRCClient
+        // register event listeners
+    }
+
+    public void start(){
+
+    }
+
+    public void stop(){
+
+    }
+}
