+++ beaconchain/src/main/java/net/consensys/beaconchain/controllers/ValidatorController.java
@@ -1,0 +1,22 @@
+package net.consensys.beaconchain.controllers;
+
+
+public class ValidatorController {
+
+    public ValidatorController(){
+
+    }
+
+    public void init(){
+        // instatiate validator
+        // register event listeners
+    }
+
+    public void start(){
+        // Begin attesting/proposing
+    }
+
+    public void stop(){
+
+    }
+}
